// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  OWNER: 'OWNER',
  PROJECT_MANAGER: 'PROJECT_MANAGER',
  DESIGNER: 'DESIGNER',
  FINANCE: 'FINANCE'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ChangeStatus = {
  DRAFT: 'DRAFT',
  PENDING_CONFIRMATION: 'PENDING_CONFIRMATION',
  CONFIRMED: 'CONFIRMED',
  REJECTED: 'REJECTED',
  WITHDRAWN: 'WITHDRAWN',
  PURCHASED: 'PURCHASED'
} as const

export type ChangeStatus = (typeof ChangeStatus)[keyof typeof ChangeStatus]


export const ConfirmationType = {
  CONFIRM: 'CONFIRM',
  REJECT: 'REJECT',
  WITHDRAW: 'WITHDRAW'
} as const

export type ConfirmationType = (typeof ConfirmationType)[keyof typeof ConfirmationType]